Location records added and fixed python part of matching

+++ tools/profileGen/scoring.py
@@ -9,17 +9,26 @@ from constants import *;
 # r := employer
 # e := employee
 def score(r, e):
+    #print("employer: ")
+    #print(r)
+    #print("employee: ")
+    #print(e)
     dist = 0
 
     for likert in LIKERTS:
             key = 'likert.%s' % (likert)
             dist += (r[key] - e[key]) ** 2
+    #print("dist")
+    #print(dist)
 
     for semDiff in SEMDIFFS:
             key = 'semDiff.%s' % (semDiff)
             dist += (r[key] - e[key]) ** 2
+    #print("dist")
+    #print(dist)
 
     maxDist = (4 ** 2) * len(LIKERTS) + (4 ** 2) * len(semDiff)
-
-    return (maxDist - dist) / maxDist
+    #print(maxDist)
+    #print((float)(maxDist - dist)/maxDist)
+    return (float)(maxDist - dist) / maxDist
